+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.8] - 2020-02-21
+### Fixed
+- BastardidentroChannel: Parse a page with meme without header
+
 ## [2.0.7] - 2020-02-21
 ### Added
 - Bastardidentro.it channel